Adding OTP + requests now are deleted after acceping an offer + change location is anow in a diffrent page to update location

+++ lib/pages/home/Customer/Accepted_reqs_tile.dart
@@ -6,12 +6,16 @@ import 'package:coffre_app/services/database.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:coffre_app/modules/users.dart';
+import 'package:internet_connection_checker/internet_connection_checker.dart';
+import 'package:overlay_support/overlay_support.dart';
 import 'package:provider/provider.dart';
 import 'package:intl/intl.dart';
 
 class acceppted_Req_Tile extends StatelessWidget {
   final AcceptedRequest acceptedRequest;
   final List acc_list;
+
+  bool hasInternet = false;
   acceppted_Req_Tile({required this.acceptedRequest, required this.acc_list});
 
   @override
@@ -37,41 +41,54 @@ class acceppted_Req_Tile extends StatelessWidget {
         margin: EdgeInsets.fromLTRB(20, 6, 20, 0),
         child: ListTile(
           trailing: TextButton.icon(
-              onPressed: () {
-                for (var item in _myAcceptedRequests) {
-                  if (item.Cust_ID == userStream.uid) {
-                    if (item.worker_ID != acceptedRequest.worker_ID) {
-                      // acc_list.add(item);
-                      DatabaseService()
-                          .AcceptenceCollection
-                          .doc(item.worker_ID)
-                          .delete() // <-- Delete
-                          .then((_) => print('Deleted'))
-                          .catchError(
-                              (error) => print('Delete failed: $error'));
-                    }
-                    if (item.worker_ID == acceptedRequest.worker_ID) {
-                      DatabaseService(uid: item.worker_ID)
-                          .updateRequestStatus(1);
+              onPressed: () async {
+                hasInternet = await InternetConnectionChecker().hasConnection;
+                if (hasInternet) {
+                  for (var item in _myAcceptedRequests) {
+                    if (item.Cust_ID == userStream.uid) {
+                      if (item.worker_ID != acceptedRequest.worker_ID) {
+                        // acc_list.add(item);
+                        DatabaseService()
+                            .AcceptenceCollection
+                            .doc(item.worker_ID)
+                            .delete() // <-- Delete
+                            .then((_) => print('Accepted'))
+                            .catchError(
+                                (error) => print('Delete failed: $error'));
+                      }
+                      if (item.worker_ID == acceptedRequest.worker_ID) {
+                        DatabaseService(uid: item.worker_ID)
+                            .updateRequestStatus(1);
+                        DatabaseService()
+                            .RequestsCollection
+                            .doc(userStream.uid)
+                            .delete() // <-- Delete
+                            .then((_) => print('Deleted'))
+                            .catchError(
+                                (error) => print('Delete failed: $error'));
 
-                      // DatabaseService().WorkingOnIt(
-                      //     item.Cust_name,
-                      //     item.Cust_ID,
-                      //     item.worker_name,
-                      //     item.worker_ID,
-                      //     DateTime.now().millisecondsSinceEpoch,
-                      //     item.price,
-                      //     1);
-                      // .then((value) => DatabaseService()
-                      //     .AcceptenceCollection
-                      //     .doc(item.worker_ID)
-                      //     .delete());
-                      Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => workingpage(item)));
+                        // DatabaseService().WorkingOnIt(
+                        //     item.Cust_name,
+                        //     item.Cust_ID,
+                        //     item.worker_name,
+                        //     item.worker_ID,
+                        //     DateTime.now().millisecondsSinceEpoch,
+                        //     item.price,
+                        //     1);
+                        // .then((value) => DatabaseService()
+                        //     .AcceptenceCollection
+                        //     .doc(item.worker_ID)
+                        //     .delete());
+                        Navigator.push(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) => workingpage(item)));
+                      }
                     }
                   }
+                } else {
+                  showSimpleNotification(Text('You have no connection!'),
+                      background: Colors.red);
                 }
               },
               icon: Icon(
@@ -101,14 +118,20 @@ class acceppted_Req_Tile extends StatelessWidget {
                     style: TextStyle(fontSize: 16, color: Colors.amber))
               ])),
           subtitle: Text(h24),
-          onTap: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => UserProfile(
-                        req: acceptedRequest,
-                      )),
-            );
+          onTap: () async {
+            hasInternet = await InternetConnectionChecker().hasConnection;
+            if (hasInternet) {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                    builder: (context) => UserProfile(
+                          req: acceptedRequest,
+                        )),
+              );
+            } else {
+              showSimpleNotification(Text('You have no connection!'),
+                  background: Colors.red);
+            }
           },
         ),
       ),
